Accept Integer args in numberPolynomials for 32-bit Julia.
On 32-bit hosts Int === Int32, so the Int64-only method never matched
and x86 CI failed with MethodError.

+++ src/multi_indices.jl
@@ -68,7 +68,7 @@ end
     `(d+n)!/(d!+n!)`
 """
 
-function numberPolynomials(d::Int64, n::Int64)
+function numberPolynomials(d::Integer, n::Integer)
     x, y = max(d, n), min(d, n)
     return UInt128(prod(UInt128(x + 1):UInt128(d + n)) ÷ factorial(UInt128(y)))
 end
